Сделал возможным получить борду и плеера из Game

+++ CodingDojo/server/src/main/java/com/codenjoy/dojo/services/multiplayer/MultiplayerServiceImpl.java
@@ -133,7 +133,7 @@ public class MultiplayerServiceImpl implements MultiplayerService{
         } else {
             PlayerGame playerGame = games.iterator().next();
             Game game = playerGame.getGame();
-            return ((Single) game).getField();
+            return game.getField();
         }
     }
